requirements.txt
# Local Voice Reception AI - Dependencies
# Python 3.11+

# === Core ===
numpy>=1.24.0,<2.0.0
scipy>=1.11.0

# === Audio Processing ===
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudio>=0.2.14

# === Speech Recognition (STT) ===
vosk>=0.3.42,<0.4.0

# === Text-to-Speech (TTS) ===
qwen-tts>=0.0.5
torch>=2.1.0
torchaudio>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0

# === LLM & RAG ===
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.22
sentence-transformers>=2.2.2
ollama>=0.1.6

# === Embedding Models ===
# intfloat/multilingual-e5-small (downloaded via HuggingFace)

# === Web UI ===
gradio>=4.19.0

# === Utilities ===
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0
rich>=13.0.0
